Allow MATLAB PAUSE function
Modified adigatorPrintTempFiles.m to allow MATLAB pause functions in
code to be differentiated.

The problem: When using Adigator with GPOPS-II and IPOPT, sometimes the
IPOPT log does not scroll continuously as one would expect. Adding a
very short pause to the Continuous.m function fixes that behavior, but
causes a ‘Cannot process statement’ error in Adigator. The change to
adigatorPrintTempFiles.m  makes Adigator ignore a line with just a
pause statement in it.

Signed-off-by: Steven R. Hall <steve@steven-hall.net>